back/core/requirements.txt
django==3.2.20
djangorestframework==3.14.0
django-cors-headers == 4.2.0
django-allauth == 0.51.0
dj-rest-auth == 2.2.5
django-filter == 23.3
python-dotenv == 1.0.0
python-decouple == 3.8
Pillow  == 10.0.1